refactor home menu activities
