+++ app/src/main/java/Activities/PatientHomeMenuActivity.java
@@ -32,39 +32,23 @@ import static GlobalSettings.GlobalSettings.PHONEEXTRA;
 import static GlobalSettings.GlobalSettings.USERIDEXTRA;
 import static android.widget.Toast.LENGTH_LONG;
 
-public class PatientHomeMenuActivity extends AppCompatActivity {
+public class PatientHomeMenuActivity extends HomeMenuActivity {
 
-    private String UserID;
-    private Patient patient = null;
-
-    private ShortCode shortCode; //for use with testing, do not include in UML
-
-    protected TextView UserIDWelcomeBox;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_patient_home_menu);
-
-        //set up text edit
-        UserIDWelcomeBox = findViewById(R.id.UserIDWelcomeBox);
-
-        //extract the user id from the intent
-        Intent intent = getIntent();
-        UserID = intent.getStringExtra(USERIDEXTRA);
-
-        //set the welcome message
-        String newText = "Welcome " + UserID;
-        UserIDWelcomeBox.setText(newText);
+    protected int getLayout() {
+        return R.layout.activity_patient_home_menu;
     }
 
-    private void FetchPatientFile() {
+    protected Class<?> getModifyActivityClass() {
+        return ModifyPatientActivity.class;
+    }
+
+    protected void FetchUserFile() {
         try {
             //get the user info for the signed in patient
             ArrayList<Patient> patients = new ElasticsearchPatientController.GetPatientTask().execute(UserID).get();
 
             //grab the first (and hopefully only) patient in the results
-            patient = patients.get(0);
+            this.user = patients.get(0);
 
         } catch (Exception e) {
             Toast.makeText(this, "Exception while fetching patient file from database",
@@ -72,47 +56,9 @@ public class PatientHomeMenuActivity extends AppCompatActivity {
         }
     }
 
-    public void onEditClick(View v) {
-        Intent intent = new Intent(this, ModifyPatientActivity.class);
-        intent.putExtra(USERIDEXTRA, UserID);
-        startActivity(intent);
-    }
-
     public void onListOfProblemsClick(View v) {
         Intent intent = new Intent(this, BrowseProblemsActivity.class);
         intent.putExtra(USERIDEXTRA, UserID);
         startActivity(intent);
-    }
-
-    public void onViewProfileClick(View v) {
-        FetchPatientFile();
-
-        Intent intent = new Intent(this, ViewUserActivity.class);
-        intent.putExtra(USERIDEXTRA, UserID);
-        intent.putExtra(EMAILEXTRA, patient.getEmail());
-        intent.putExtra(PHONEEXTRA, patient.getPhoneNumber());
-        startActivity(intent);
-    }
-
-    public void onDeleteClick(View v) {
-        //create fragment that comes up and asks if the user is sure
-    }
-
-    //generate login code for another device to use for login using security token
-    public void onGenerateCodeClick(View v) {
-        try {
-            this.shortCode = ShortCodeController.AddCode(this.UserID, this);
-            Toast.makeText(this,
-                    "Added code " + shortCode.getShortSecurityCode(), LENGTH_LONG).show();
-        } catch(failedToFetchSecurityTokenException e) {
-            Toast.makeText(this,
-                    "Unable to load a security token for this user", LENGTH_LONG).show();
-        } catch(failedToAddShortCodeException e) {
-            Toast.makeText(this, "Unable to add the short code", LENGTH_LONG).show();
-        }
-    }
-
-    public ShortCode getShortCode() {
-        return shortCode;
     }
 }
